requirements.txt
contourpy==1.2.1
cycler==0.12.1
enum34==1.1.10
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.9.0
netifaces==0.11.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
psutil==5.9.8
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
pyvalid==1.0.4
shiboken6==6.7.0
six==1.16.0
